Remove trait and move to abstract class
Remove code duplications
